Extract Crowding model
