+++ assets/src/realtime.d.ts
@@ -18,20 +18,6 @@ export interface BlockWaiver {
   remark: string | null
 }
 
-export type OccupancyStatus =
-  | "NO_DATA"
-  | "EMPTY"
-  | "MANY_SEATS_AVAILABLE"
-  | "FEW_SEATS_AVAILABLE"
-  | "FULL"
-
-export interface Crowding {
-  load: number | null
-  capacity: number | null
-  occupancyStatus: OccupancyStatus
-  occupancyPercentage: number | null
-}
-
 export interface DataDiscrepancy {
   attribute: string
   sources: DataDiscrepancySource[]
